restructured redux files
