+++ browser/react/components/navbar/NavbarContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import Navbar from './Navbar';
-import { leaveProject } from '../../redux/selectedProject'
+import { leaveProject } from '../../store/modules/selectedProject'
 
 const mapDispatchToProps = dispatch => ({
   removeProject: () => dispatch(leaveProject())
