+++ browser/react/store/modules/index.js
