+++ browser/react/store/modules/projects.js
