+++ browser/react/store/modules/selectedProject.js
